requirements.txt
requests==2.32.3
beautifulsoup4==4.12.2
python-dateutil==2.8.2